tu_proyecto/requirements.txt
Flask
Flask-Mail
celery[redis]
redis
gunicorn
python-dotenv